+++ src/Fp.Fs/FormatSingleProcessorFsWrapper.cs
@@ -22,16 +22,20 @@ public class FormatSingleProcessorFsWrapper : FsProcessor
     /// <inheritdoc />
     protected override IEnumerable<Data> ProcessSegmentedImpl()
     {
+        _baseProcessor.Info = Info;
         var input = OpenMainFile();
         _baseProcessor.Prepare(input, Name, Configuration);
-        return _baseProcessor.TryProcess(out Data? data) ? new[] { data! } : Array.Empty<Data>();
+        return _baseProcessor.TryProcess(out Data? data)
+            // ReSharper disable once RedundantSuppressNullableWarningExpression
+            ? new[] { data! }
+            : Array.Empty<Data>();
     }
 
     /// <inheritdoc />
     public override void Cleanup(bool warn = false)
     {
         base.Cleanup(warn);
-        _baseProcessor.Cleanup();
+        _baseProcessor.Cleanup(warn);
     }
 }